**Add .gitignore to exclude temporary files**
This commit introduces a new `.gitignore` file to the repository to ensure that specific temporary files generated during the build process are not tracked by Git. The following entries have been added:

- `/BACKEND/JAVA/RetoTecnicoSofka25Diciembre2025Backend/tempJar2`
- `/BACKEND/JAVA/RetoTecnicoSofka25Diciembre2025Backend/tempSwagger`

These entries will prevent clutter in the repository by ignoring temporary JAR files and Swagger-related files, which are not necessary for version control. This helps maintain a cleaner project history and reduces the risk of accidentally committing unnecessary files.