Add JpaAccountWriteRepositoryAdapter for account persistence
This commit introduces the JpaAccountWriteRepositoryAdapter class, which implements the AccountWriteRepositoryPort interface to facilitate account persistence in the application. The adapter uses an instance of AccountWriteJpaRepository to handle the saving of Account entities. This addition establishes a necessary component for writing account data to the database, promoting adherence to the Ports and Adapters architecture pattern. The class is annotated with @Component to allow it to be automatically detected by Spring's component scanning, ensuring that it can be injected wherever needed in the application.