Add GetAccountQuery DTO for account retrieval
Introduced a new Data Transfer Object (DTO) called GetAccountQuery. This class is designed to facilitate the fetching of a single account by its identifier. It includes a constructor that takes an account ID as a parameter and a getter method to retrieve that ID. This addition enhances the API's capabilities by allowing more structured queries for account data.