Add MovementEventReadJpaRepository interface for reading movement events
This commit introduces a new interface, MovementEventReadJpaRepository, that extends JpaRepository for accessing MovementEvent entities in the persistence layer. The repository is annotated with @Repository and includes a method, findByPublishedFalseOrderByCreatedAtAsc, which retrieves a list of movement events that have not been published, ordered by their creation date in ascending order. This addition will facilitate easier querying of movement events within the application and improve overall data management related to financial transactions.