Add MovementEventWriteJpaRepository
This commit introduces a new interface, MovementEventWriteJpaRepository, which extends JpaRepository for the MovementEvent entity. The repository is annotated with @Repository, allowing it to be recognized by Spring as a bean for data access operations. This addition facilitates CRUD operations for MovementEvent instances in the underlying database, streamlining data management within the application.